% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashed_soc_to_int.R
\name{dashed_soc_to_int}
\alias{dashed_soc_to_int}
\title{Utility for converting xx-xxxx SOC codes to 6-digit integers}
\usage{
dashed_soc_to_int(.dashed)
}
\arguments{
\item{.dashed}{a string vector of SOC codes}
}
\value{
an integer vector of 6-digit SOC codes
}
\description{
Utility for converting xx-xxxx SOC codes to 6-digit integers
}
