% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EDUCATION_ASSIGNMENTS}
\alias{EDUCATION_ASSIGNMENTS}
\title{Education and training assignments by detailed occupation}
\format{
A six-column table.

\describe{
\item{Group}{The level of the hierarchy an occupation type belongs to.}
\item{Major}{Which top-level group this occupation belongs to}
\item{Minor}{Which 2nd-level group this occupation belongs to}
\item{Broad}{Which 3rd-level group this occupation belongs to}
\item{Detailed}{The SOC code for this occupation type or group of types.}
\item{Title}{A verbal description of the occupation type or group.}
}
}
\source{
\url{https://www.bls.gov/emp/ind-occ-matrix/education.xlsx}

\url{https://www.bls.gov/emp/tables/education-and-training-by-occupation.htm}
}
\usage{
EDUCATION_ASSIGNMENTS
}
\description{
A table of educational and training requirements for many different jobs.
}
\seealso{
\code{\link[=parse_education_assignments_file]{parse_education_assignments_file()}}
}
\keyword{datasets}
