% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EDUCATIONAL_ATTAINMENT_LEVELS}
\alias{EDUCATIONAL_ATTAINMENT_LEVELS}
\title{Levels of educational attainment, according to the BLS}
\format{
A two-column table.

\describe{
\item{credential}{A verbal description of a level of education.}
\item{bls_id}{A two-digit integer}
}
}
\source{
\url{https://www.bls.gov/emp/tables/education-and-training-by-occupation.htm}
}
\usage{
EDUCATIONAL_ATTAINMENT_LEVELS
}
\description{
A table of educational levels with two-digit coedes
}
\seealso{
\link{EDUCATION_ASSIGNMENTS}
}
\keyword{datasets}
