% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLOSSARY_OF_AREA_TYPES}
\alias{GLOSSARY_OF_AREA_TYPES}
\title{Types of geographic area}
\format{
An object of class \code{tibble/tbl_df/tbl/data.frame}
with 14 rows and 2 columns
\describe{
\item{area_code_type}{\verb{<chr>} upper-case letters}
\item{areatype_text}{\verb{<chr>} a short description of the kind of area}
}
}
\source{
\url{https://download.bls.gov/pub/time.series/la/la.area_type}
}
\usage{
GLOSSARY_OF_AREA_TYPES
}
\description{
A look-up table of one-letter codes for specifying the kind of geographic
area described by a BLS series.
}
\keyword{datasets}
