% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLOSSARY_OF_AREAS}
\alias{GLOSSARY_OF_AREAS}
\title{Specific geographic areas}
\format{
An object of class \code{tibble/tbl_df/tbl/data.frame}
with 8325 rows and 4 columns
\describe{
\item{area_type_code}{\verb{<chr>} a single-letter code}
\item{area_code     }{\verb{<chr>} a fifteen-character code, basically two characters plus FIPS}
\item{area_text     }{\verb{<chr>} a short description of the }
\item{sort_sequence }{\verb{<int>} the sort order by FIPS code}
}
}
\source{
\url{https://download.bls.gov/pub/time.series/la/la.area}
}
\usage{
GLOSSARY_OF_AREAS
}
\description{
A look-up table of fifteen-character codes for the specific geographic areas
that can be described by a BLS series. For example, "CN2004900000000" is
Elk County, Kansas, and "CT2718116000000" is Eden Prairie, Minnesota.
}
\keyword{datasets}
