% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLOSSARY_OF_SEASONAL_ADJUSTMENTS}
\alias{GLOSSARY_OF_SEASONAL_ADJUSTMENTS}
\title{Types of seasonal adjustments}
\format{
An object of class \code{tibble/tbl_df/tbl/data.frame}
with 2 rows and 2 columns
\describe{
\item{adjustment_code}{\verb{<chr>} "U" or "S"}
\item{adjustment_text}{\verb{<chr>} "Unadjusted" or "Seasonally Adjusted"}
}
}
\source{
\url{https://www.bls.gov/help/hlpseas.htm}
}
\usage{
GLOSSARY_OF_SEASONAL_ADJUSTMENTS
}
\description{
A look-up table of one-letter codes for specifying whether or not a series'
values are seasonally adjusted.
}
\keyword{datasets}
