% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLOSSARY_OF_SURVEYS}
\alias{GLOSSARY_OF_SURVEYS}
\title{The menu of surveys at the BLS API}
\format{
An object of class \code{tibble/tbl_df/tbl/data.frame}
with 2 rows and 2 columns
\describe{
\item{Prefix}{\verb{<chr>} a two-letter code that starts every series associated with this survey}
\item{Survey}{\verb{<chr>} A short title for the survey.}
}
}
\source{
\url{https://download.bls.gov/pub/time.series/overview.txt}

\url{https://api.bls.gov/publicAPI/v2/surveys}
}
\usage{
GLOSSARY_OF_SURVEYS
}
\description{
Each of the surveys has an official two-letter prefix code.
}
\keyword{datasets}
