% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hercbls-package.R
\docType{package}
\name{hercbls-package}
\alias{hercbls}
\alias{hercbls-package}
\title{hercbls: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://higherx4racine.github.io/hercbls/}
}

}
\author{
\strong{Maintainer}: Ben Taft \email{ben.taft@career2cradle.org}

Other contributors:
\itemize{
  \item Higher Expectations for Racine County [copyright holder]
}

}
\keyword{internal}
