% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{METADATA_FOR_WP}
\alias{METADATA_FOR_WP}
\title{Lookup tables for codes used by series "WP."}
\format{
a list of five data frames
\subsection{footnote}{

An object of class \code{spec_tbl_df/tbl_df/tbl/data.frame}
with 2 rows and 2 columns

\describe{
\item{footnote_code}{\verb{<chr>} a one-letter code}
\item{footnote_text}{\verb{<chr>} an explanation of the footnote}
}
}

\subsection{group}{

An object of class \code{spec_tbl_df/tbl_df/tbl/data.frame}
with 56 rows and 2 columns (S3: spec_tbl_df/tbl_df/tbl/data.frame)

\describe{
\item{group_code}{\verb{<chr>} a two-character code}
\item{group_text}{\verb{<chr>} a broad category of commodities}
}
}

\subsection{item}{

An object of class \code{spec_tbl_df/tbl_df/tbl/data.frame}
with 4168 rows and 4 columns

\describe{
\item{group_code}{\verb{<chr>} a two-character code}
\item{item_code}{\verb{<chr>} a one- to six-digit code for a specific commodity}
\item{item_name}{\verb{<chr>} the name of the commodity}
}
}

\subsection{period}{

An object of class \code{spec_tbl_df/tbl_df/tbl/data.frame}
with 13 rows and 3 columns

\describe{
\item{period}{\verb{<chr>} a three-character code like "M01" or "M13"}
\item{period_abbr}{\verb{<chr>} short text for the period, like "JAN" or "AN AV"}
\item{period_name}{\verb{<chr>} the full name of the period}
}
}

\subsection{seasonal}{

An object of class \code{spec_tbl_df/tbl_df/tbl/data.frame}
with 2 rows and 2 columns

\describe{
\item{seasonal_code}{\verb{<chr>} either "U" or "S"}
\item{seasonal_text}{\verb{<chr>} either "Unadjusted" or "Seasonally Adjusted."}
}
}
}
\source{
\url{https://download.bls.gov/pub/time.series/wp/wp.txt}
}
\usage{
METADATA_FOR_WP
}
\description{
\subsection{METADATA_FOR_WP}{
}
}
\keyword{datasets}
