% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_education_assignments_file.R
\name{parse_education_assignments_file}
\alias{parse_education_assignments_file}
\title{Read a spreadsheet of S.O.C. codes and educational levels.}
\usage{
parse_education_assignments_file(.ed_file)
}
\arguments{
\item{.ed_file}{the path to the file downloaded from the BLS.}
}
\value{
a data table with 5 columns
\describe{
\item{Title}{The name of the occupation type or group.}
\item{soc_code}{The \link{SOC_HIERARCHY} code for the occupation}
\item{Education}{A text description of the education needed for the job}
\item{Experience}{Text stating previous work history needed for the job}
\item{Training}{How much on-the-job training is typical for the job}
}
}
\description{
This is an internal function for creating the \link{EDUCATION_ASSIGNMENTS} data
table.
}
\seealso{
\link{EDUCATION_ASSIGNMENTS}
}
