% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_oes_download.R
\name{parse_oes_download}
\alias{parse_oes_download}
\title{Read an excel spreadsheet pulled from the BLS site.}
\usage{
parse_oes_download(.fullpath)
}
\arguments{
\item{.fullpath}{the location and file name where you saved the data}
}
\value{
a data frame with at least three columns
\describe{
\item{title}{The verbal description of the occupation or group}
\item{soc_code}{The six-digit integer code for the occupation or group}
\item{...}{one or more columns of data, depending on what you downloaded}
}
}
\description{
One way to get wage data from the BLS is to use
\href{https://data.bls.gov/oes/#/home}{this form}
}
