% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_soc_hierarchy_file.R
\name{parse_soc_hierarchy_file}
\alias{parse_soc_hierarchy_file}
\title{Read a spreadsheet of Standard Occupational Classification codes}
\usage{
parse_soc_hierarchy_file(.hierarchy_file)
}
\arguments{
\item{.hierarchy_file}{the full path to the spreadsheet}
}
\value{
a data frame of classified occupation types
}
\description{
Read a spreadsheet of Standard Occupational Classification codes
}
