% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{METADATA_FOR_LA}
\alias{METADATA_FOR_LA}
\title{Lookup tables for codes used by series "LA."}
\format{
a list of seven data frames
\subsection{area}{

An object of class \code{spec_tbl_df/tbl_df/tbl/data.frame}
with 8,325 rows and 4 columns

\describe{
\item{area_type_code}{\verb{<chr>} a one-letter code}
\item{area_code}{\verb{<chr>} the type_code followed by a FIPS code}
\item{area_text}{\verb{<chr>} a short description of the area, usually its name}
\item{sort_sequence}{\verb{<int>} which row this area would be if they were sorted by FIPS codes}
}
}

\subsection{area_type}{

An object of class \code{spec_tbl_df/tbl_df/tbl/data.frame}
with 14 rows and 2 columns

\describe{
\item{area_type_code}{\verb{<chr>} a one-letter code}
\item{areatype_text}{\verb{<chr>} a short description of the type of area }
}
}

\subsection{footnote}{

An object of class \code{spec_tbl_df/tbl_df/tbl/data.frame}
with 7 rows and 2 columns

\describe{
\item{footnote_code}{\verb{<chr>} a one-letter code}
\item{footnote_text}{\verb{<chr>} an explanation of the footnote}
}
}

\subsection{measure}{

An object of class \code{spec_tbl_df/tbl_df/tbl/data.frame}
with 7 rows and 2 columns (S3: spec_tbl_df/tbl_df/tbl/data.frame)

\describe{
\item{measure_code}{\verb{<chr>} a two-character code}
\item{measure_text}{\verb{<chr>} a description of what the values measure}
}
}

\subsection{period}{

An object of class \code{spec_tbl_df/tbl_df/tbl/data.frame}
with 13 rows and 3 columns

\describe{
\item{period}{\verb{<chr>} a three-character code like "M01" or "M13"}
\item{period_abbr}{\verb{<chr>} short text for the period, like "JAN" or "AN AV"}
\item{period_name}{\verb{<chr>} the full name of the period}
}
}

\subsection{seasonal}{

An object of class \code{spec_tbl_df/tbl_df/tbl/data.frame}
with 2 rows and 2 columns

\describe{
\item{seasonal_code}{\verb{<chr>} either "U" or "S"}
\item{seasonal_text}{\verb{<chr>} either "Unadjusted" or "Seasonally Adjusted."}
}
}

\subsection{state_region_division}{

An object of class \code{spec_tbl_df/tbl_df/tbl/data.frame}
with 53 rows and 2 columns

\describe{
\item{srd_code}{\verb{<chr>} a two-digit code}
\item{srd_text}{\verb{<chr>} the name of the area}
}
}
}
\source{
\url{https://download.bls.gov/pub/time.series/la/la.txt}
}
\usage{
METADATA_FOR_LA
}
\description{
\subsection{METADATA_FOR_LA}{
}
}
\keyword{datasets}
